+++ heat/core/tests/test_manipulations.py
@@ -3276,13 +3276,21 @@ class TestManipulations(TestCase):
         self.assertTrue((np_tiled == ht_tiled.numpy()).all())
         self.assertTrue(ht_tiled.dtype is x.dtype)
 
-        # test scalar x
+        # test scalar DNDarray x
         x = ht.array(9.0)
         reps = (2, 1)
         ht_tiled = ht.tile(x, reps)
         np_tiled = np.tile(x.numpy(), reps)
         self.assertTrue((np_tiled == ht_tiled.numpy()).all())
         self.assertTrue(ht_tiled.dtype is x.dtype)
+
+        # test scalar x
+        x = 10
+        reps = (2, 1)
+        ht_tiled = ht.tile(x, reps)
+        np_tiled = np.tile(x, reps)
+        self.assertTrue((np_tiled == ht_tiled.numpy()).all())
+        self.assertTrue(ht_tiled.dtype is ht.int64)
 
         # test distributed tile along split axis
         # len(reps) > x.ndim